logic of parser and admin view splited to two servlets
