+++ web/src/main/java/com/infoshareacademy/dao/RecipeRepositoryDao.java
@@ -3,15 +3,14 @@ package com.infoshareacademy.dao;
 
 import com.infoshareacademy.domain.Recipe;
 import com.infoshareacademy.repository.RecipeRepository;
-import java.util.List;
+
 import javax.ejb.Stateless;
+import java.util.List;
 
 @Stateless
 public class RecipeRepositoryDao {
 
   public void loadRepository(List<Recipe> recipesList) {
-    if (RecipeRepository.getRecipesList().isEmpty()) {
       RecipeRepository.getRecipesList().addAll(recipesList);
-    }
   }
 }
