+++ web/src/main/java/com/infoshareacademy/freemarker/TemplateProvider.java
@@ -2,10 +2,11 @@ package com.infoshareacademy.freemarker;
 
 import freemarker.template.Configuration;
 import freemarker.template.Template;
-import java.io.IOException;
+
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 import javax.servlet.ServletContext;
+import java.io.IOException;
 
 @RequestScoped
 public class TemplateProvider {
