+++ web/src/main/java/com/infoshareacademy/repository/RecipeRepository.java
@@ -1,6 +1,7 @@
 package com.infoshareacademy.repository;
 
 import com.infoshareacademy.domain.Recipe;
+
 import java.util.ArrayList;
 import java.util.List;
