+++ web/src/main/java/com/infoshareacademy/service/LoadParsedDataToDao.java
@@ -2,9 +2,10 @@ package com.infoshareacademy.service;
 
 import com.infoshareacademy.dao.RecipeRepositoryDao;
 import com.infoshareacademy.domain.Recipe;
-import java.util.List;
+
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
+import java.util.List;
 
 @RequestScoped
 public class LoadParsedDataToDao {
