--- web/src/main/webapp/WEB-INF/fm-templates/admin-upload-data.ftlh
@@ -1,9 +1,0 @@
-<div class="btn btn-secondary btn-lg active" role="button" aria-pressed="true">
-  <form id="form-upload">
-    <button type="button" class="btn btn-secondary" onclick=setStatus()>Upload</button>
-  </form>
-</div>
-<script>
-  function setStatus() {
-  }
-</script>
