+++ web/src/main/webapp/WEB-INF/fm-templates/admin-view.ftlh
@@ -19,9 +19,11 @@
     <h1 class="display-3">Admin Panel</h1>
   </div>
 </div>
-<div>
-    <#include "admin-upload-data.ftlh">
-</div>
+  <div class="btn btn-secondary btn-lg active" role="button" aria-pressed="true">
+    <form id="form-upload" action="/upload-data" method="post">
+      <button type="submit" class="btn btn-secondary">Upload</button>
+    </form>
+  </div>
 <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js"></script>
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
